Move custom cluster definitions and costants to __init__.py.
